GYM-11 change the entities and also add the return of forus with limi…

+++ src/main/java/org/filrouge/gymcommunity/dto/userNutr/UserNutritionResDTO.java
@@ -1,8 +1,8 @@
 package org.filrouge.gymcommunity.dto.userNutr;
 
 public record UserNutritionResDTO(
-        float dailyCalorieGoal,
-        float proteins,
-        float carbs,
-        float fats
+        long dailyCalorieGoal,
+        long proteins,
+        long carbs,
+        long fats
 ) {}
